+++ CommonConcepts/CommonConcepts.Test/Helpers/Program.cs
@@ -40,10 +40,10 @@ namespace CommonConcepts.Test
         {
             return new RhetosHostBuilder()
                 .ConfigureRhetosHostDefaults()
-                .ConfigureConfiguration(builder =>
-                {
-                    builder.AddJsonFile(RhetosAppEnvironment.LocalConfigurationFileName);
-                })
+                .ConfigureConfiguration(builder => builder
+                    .AddJsonFile("rhetos-app.settings.json")
+                    .AddJsonFile("rhetos-app.local.settings.json")
+                )
                 .ConfigureContainer(builder =>
                 {
                     builder.RegisterType<ProcessUserInfo>().As<IUserInfo>();