+++ ChangeLog.md
@@ -48,6 +48,9 @@
     * Code that generates or reads assets files should use IAssetsOptions.AssetsFolder instead.
     * Code that uses Paths class for Rhetos components initialization (for example SearchForAssembly) should be refactored to use RhetosHost class instead.
     * Code that uses the ResourcesFolder should use  IAssetsOptions.AssetsFolder instead.
+22. Removed RhetosAppEnvironment class and AddRhetosAppEnvironment method.
+    * Configuration files "rhetos-app.settings.json" and "rhetos-app.local.settings.json" are no longer automatically loaded. If you want to use the same files from Rhetos v4, load them in your application in Program.CreateRhetosHostBuilder method:
+      `return new RhetosHostBuilder().ConfigureRhetosHostDefaults().ConfigureConfiguration(builder => builder.AddJsonFile("rhetos-app.settings.json").AddJsonFile("rhetos-app.local.settings.json"));`
 
 ## 4.3.0 (TO BE RELEASED)